requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pyodbc==5.0.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
Werkzeug